Fix contradiction picked up by Clang warning of string::find's failure code.
Per spec, size_t is unsigned, and hence can never be negative, so
end < 0 is a contradiction. The return value of string::find when it
fails to find anything is the maximum value of size_t, which is also
the value of string::npos, which should be used for clarity's sake.