Change {{VERSION}} placeholder to the current version 0.9.1.
Initially this was set via shell script during the export process from Google's
repository to the open-source one, but now that most changes are coming from the
open-source repository, it doesn't make sense.  Since it's only in 2 places and
changes will likely be infrequent, it's easier to update it manually.