Added next/prev fields to GumboNode.
This technically is an API addition and should normally include a bump of the
minor version number, but since we're on pre-release (0.9), we may just do it by
bumping the release number.  Or we could release a 0.10; that may be more fair
to clients and honor our promise of semantic versioning more tightly; that may
be worth it, even though a version 0.10 looks a bit weird.