Match elements on qualified names, not just local names
There are many places where the HTML spec says to look for an HTML
element (i.e., an element in the HTML namespace) with a given tag, but
where we were only looking at tag names and ignoring the namespace name.

Now we always use qualified names where required. This is implemented
using a new GumboQualName type, which is just a bit field that combines
a GumboNamespaceEnum with a GumboTag. A series of macros make it easy to
construct and inspect these bit fields. It's currently represented by a
uintptr_t. This is larger than necessary; only 9 bits are required. At
first I attempted to use a `short` but the compiler didn't like that
being used with varargs, so then I tried an `unsigned int` but the
compiler didn't like casting that to a pointer, so here we are. I also
tried `typedef enum { QNSIZE = SHORT_MAX } GumboQualName` to induce more
compiler warnings when mistakenly passing a tag to a function that
expects a qualified name. This worked nicely but generated warnings
about missing cases in switch statements. I'm not sure what the best
option is, though I am quite tempted by the extra warnings the enum
provides.