tags: Use a perfect hash for lookups
The previous version using `strcasecmp` over an array was a bottleneck
on the library.

This version uses a simple, minimal perfect hash table (computed via
`mph`) to convert tag names into strings. Since we're now hashing tag
names, we can pass in the length of the tag name explicitly, and avoid
the superfluous allocations that the tokenizer was performing in order
to NULL-terminate the tag. This is implemented on the new
`gumbo_tagn_enum` API.

The old `gumbo_tag_enum` API has been left as a thin wrapper to keep
backwards compatibility -- it is not used internally by the library.

`mph` was chosen for the perfect hash function because it generates
hashes that are slightly slower than GPerf but significantly simpler,
and occuppying an order of magnitude less memory (as they don't
need a full copy of all the strings in the set for hashing).

If the tag lookup function proves to be a bottleneck, this decision can
be re-evaluated in the future.