parser: Simplify the `element_in_specific_scope` calls
The old implementation using 2 tagsets was being rather wasteful with
stack space, allocating 2 whole sets when one of them always contains a
single tag element. Knowing that the `expected` elements must always be
in the HTML namespace, we can simplify these APIs by passing an array of
elements and stop allocating so much space on the stack.