Change repeated calls to string_buffer_destroy/string_buffer_init to a string_buffer_clear that just zeros out the vector instead of reallocating.  Benchmarks:
    parse_time: median=3217.00, 95th%=13199.00
    traversal_time: median=51000.00, 95th%=343000.00
    allocations: median=12459.50, 95th%=49388.75
    bytes_allocated: median=441710.00, 95th%=1756603.10
    high_water_mark: median=194527.00, 95th%=796959.05

    hacker_news.html: 4574 microseconds.
    html5_spec.html: 654969 microseconds.
    xinhua.html: 19776 microseconds.
    baidu.html: 5675 microseconds.
    yahoo.html: 27398 microseconds.
    google.html: 30849 microseconds.
    bbc.html: 7762 microseconds.
    arabic_newspapers.html: 5798 microseconds.
    wikipedia.html: 35138 microseconds.

This gives a roughly 8-10% improvement in CPU performance, and reduces the total
number of bytes allocated by about 5%.  However, the cost of it is that it
eliminates virtually all the memory savings of changing the initial vector
numbers (it's maybe 5% over master, vs. about 30%).  The source of that is that
now the temporary vectors can grow unbounded, and so a large string will
continue consuming memory until parsing is complete.  I'm investigating some
approaches to mitigate that.