tags: Use a gperf-based hash table
The original perfect hash table implementation was generated using
`mph`, a lesser known hash table generator, which was manually modified
by me to generate output that suits better the use case in Gumbo Parser.

The hash function, although simple, had several issues. As @nostrademons
pointed out, having the tag generator sitting in a random repository
makes it hard for other developers to build the tag sets. Furthermore,
the hash function itself was not particularly efficient, as it hashed
the whole string and performed several lookups from two separate tables.

This commit changes the hash table implementation to use GNU gperf,
widely available in all Linux distributions and Mac OS X. The issue with
gperf, however, is that the generated hash tables make very inefficient
use of memory, by duplicating the whole string key pool (another concern
of @nostrademons when the issue of perfect hashing was initially
raised)

Because of this, an accompanying script, `genperf.py` has been added to
the repository. This script turns the output of gperf (tested with all
2.x and 3.x versions) into an optimal representation of a tagset that
compiles direcly in the library with no further modification.

The generated hash table is 400% smaller than the previous implementation
and up to 40% faster, and can be regenerated trivially in all platforms
just by having gperf as a depedency. The script simply converts the
string array generated by gperf into an array of GumboTag elements, that
fit in a byte each (hence the significant reduction in memory usage).

Conflicts:
	src/tag.c