tags: Fix issues in the perfect hash implementation
The result of the perfect hash for tag names was not being properly
compared against the candidate string, as no length checks were being
taken into account.

Hence, looking up `t` would match `textarea` because both strings hashed
to the same value (by coincidence), and yet only the first byte of the
strings was being compared for a full match.

The issue has been fixed by adding a table of tag lengths, which will
additionally speed up the rejection of invalid strings.

In order to simplify the generation of the several automated Tag tables,
and to remove the depedency on `sed` from the Makefile, a simple
`gentags.py` has been added to the codebase.

Running `make gentags` will generate all the tables **and** the perfect
hash function, assuming that Python and the correct version of MPH is in
the path.

An updated version of mph has been pushed to the following repository:

	https://github.com/vmg/mph

It contains all the changes required to generate case-insensitive hashes
just like the ones used in the library, with no further modification to
the hash output.

Conflicts:
	src/tag.c